Refactor pickup by renaming and changing file structure
