+++ client/spa-pages/index.ts
@@ -1,4 +1,4 @@
 export * from "./components/HomePage";
-export * from "./components/HomePickupPage";
+export * from "./components/PickupPage";
 export * from "./components/InstructionsPage";
 export * from "./components/MapPage";
